Add files via upload
**#Global internet usage**
**## This dashboard represents the usage of the internet by the countries globally and their counts. The measures are based on the income per person, Internet user rate, and urban rate. **

In this report, I have used DAX and aggregate functions for the sum of totals, RANKX, and TOPN filter functions. Using the  PowerBI RANKX , all the countries have been ranked based on their max income per person, urban rate & user rate. 
Using the TOPN, the top 10 countries have been filtered based on their income per person. Monaco stands to the highest in making income. Iceland stands to be in the number 1 position on the basis of internet user rate. 
The symbolic arrows in the tabular chart in red, yellow, and green categorize the countries as above 30%, 60%, and 90% respectively based on their income, user rate, and urban rate.